pyproject.toml
[project]
name = "clinical-notes-icd-10-finetuning"
version = "2.0.0"
description = "MIMIC-IV clinical notes and ICD-10 diagnosis code extraction and processing pipeline"
readme = "icd10_fine_tune/README.md"
requires-python = ">=3.9"
authors = [
    {name = "RhythmX AI Team"}
]
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "tenacity>=8.2.0",
    "pyarrow>=12.0.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.3.0",
    "openai>=1.0.0",
    "requests>=2.31.0",
    "sentence-transformers>=3.0.0",
    "scikit-learn>=1.3.0",
    "tf-keras>=2.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.5.0",
]

# Fine-tuning dependencies for ICD-10 code extraction
# Install with: uv pip install -e ".[finetune]"
#
# EDUCATIONAL NOTE - Fine-Tuning Stack:
# - unsloth: Memory-optimized fine-tuning (2x faster, 50% less VRAM)
# - peft: Parameter-Efficient Fine-Tuning (LoRA implementation)
# - trl: Transformer Reinforcement Learning (SFTTrainer)
# - bitsandbytes: 4-bit/8-bit quantization for VRAM reduction
# - accelerate: Multi-GPU training support (DDP)
# - evidently: Model evaluation and monitoring dashboard
finetune = [
    "transformers>=4.41.0",
    "peft>=0.11.0",
    "trl>=0.9.0",
    "accelerate>=0.30.0",
    "bitsandbytes>=0.43.0",
    "datasets>=2.19.0",
    "evidently>=0.4.20",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["MIMIC_VI_Data_Extract"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
